Modify kthread-based server to CMWQ-based server
The kthread-based version requires creating a new kthread for each
received request, which consumes a significant amount of time in
kthread creation. However, by using CMWQ, we only need to enqueue
the tasks into the queue. Moreover, a single kthread in CMWQ may
be able to handle multiple tasks that are placed in the queue,
resulting in better cache locality.